refactor: separating AuthContext appropriately
