+++ src/lib/contexts/AuthContext.ts
@@ -1,0 +1,14 @@
+import { User } from "firebase/auth";
+import { createContext } from "react";
+
+interface AuthContextType {
+  user: User | null;
+  loading: boolean;
+  token: string | null;
+  refreshToken: () => Promise<void>;
+  loginWithToken: (user: User) => Promise<void>;
+}
+
+export const AuthContext = createContext<AuthContextType | undefined>(
+  undefined
+);
