+++ src/lib/hooks/useAuth.ts
@@ -1,0 +1,12 @@
+"use client";
+
+import { AuthContext } from "@/lib/contexts/AuthContext";
+import { useContext } from "react";
+
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+  if (context === undefined) {
+    throw new Error("useAuth must be used within an AuthProvider");
+  }
+  return context;
+};
